// next-js-training/src/libs/schemas/loginSchema.ts
import { z } from "zod";

export const loginSchema = z.object({
    email: z.string().email({
        message: "Please enter your email"
    }),
    password: z.string().min(6, {
        message: "Password must be at least 6 characters"
    })
})

export type LoginSchema = z.infer<typeof loginSchema>